requirements.txt
torch == 2.1.1
torchvision == 0.16.1
xformers
albumentations==1.3.1

numpy == 1.26.2
scipy == 1.11.4
scikit-learn == 1.3.2
pandas == 2.1.4
matplotlib == 3.8.2
seaborn == 0.13.0

hydra-core==1.3.2
omegaconf==2.3.0

tqdm
jupyter == 1.0.0
tensorboard
